ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-dotenv==1.0.1
google-generativeai==0.8.0
python-multipart==0.0.9

# AI/ML Libraries
tensorflow==2.17.0
torch==2.4.1
transformers==4.44.2
sentence-transformers==3.1.0
spacy==3.8.2
nltk==3.9.1
textstat==0.7.4

# Computer Vision
opencv-python==4.10.0.84
Pillow==10.4.0
pytesseract==0.3.13

# Audio Processing
librosa==0.10.2.post1
soundfile==0.12.1
speechrecognition==3.10.4

# Web3/Blockchain
web3==7.2.0
eth-account==0.13.3

# Data Processing
numpy==2.1.1
pandas==2.2.2

# HTTP Clients
httpx==0.27.2
aiohttp==3.10.5
requests==2.32.3

# Caching & Queue
redis==5.0.8
celery==5.4.0

# Database
pymongo==4.8.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.34

# Utilities
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
